Fix w2sidebar user interactions on windows.
Each node of the w2sidebar has to have a unique id attribute.
Before, we used the path of the file system object it represents.
As it turns out, these ids cannot have an arbitrary format.
Especially windows file system paths caused trouble.
This resulted in a failure when using the ids to lookup the node.