Adding a fresh filename generator to be used in the creation of new multi-models.
When a user creates multi-models the name suggested in the dialog box is the first non-existing file in the range: new-mm, new-mm1, ..., new-mm99. In case all the 100 default names are already used the user must manually insert a new name in the dialog box.

The same holds for multi-models created from existing SysML configuration files, but in that case the range is: $name, $name1, ..., $name99 where $name is the SysML configuration filename from which the model is being created.